+++ codes/__init__.py
@@ -1,6 +1,5 @@
-# import .py
+# import file
 from codes import lib
-from codes import load_textures
 
 # import folders
 from codes import blocks
